New year modification

+++ Include/NSOBochs.h
@@ -5,8 +5,8 @@
 
 u8 kernel_printCharBOCHS(const char);
 u8 kernel_printStringBOCHS(const char*);
-u8 kernel_printUNumberBOCHS(const u64);
-u8 kernel_printHexBOCHS(const u64);
+u8 kernel_printUNumberBOCHS(const u32);
+u8 kernel_printHexBOCHS(const u32);
 u8 kernel_printBoolBOCHS(const u8);
 u8 kernel_printfBOCHS(const char*, ...);
