+++ Source/utils/NSOStringUtils.c
@@ -8,10 +8,10 @@
  * string: The NULL-terminated string to be calculated
  *
  *
- * Return: Returns a unsigned 64 bit number representing the length of the string
+ * Return: Returns a unsigned 32 bit number representing the length of the string
  */
-u64 kernel_stringLength(const char* string){
-	u64 length = 0;
+u32 kernel_stringLength(const char* string){
+	u32 length = 0;
 	char* temp = (char*)string;
 
 	if (string != NULL){
